refactor(lint-cli): hoist hybrid status out of the lint runner
The factory writes the hybrid status during config evaluation and the lint
runner reads it, so it is genuinely shared rather than a lint-cli internal.
Moving it to src/ removes factory.ts's reach into ../lint-cli/ and keeps the
runner's tree (yargs, concurrently) out of dist/index.mjs.
